requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
file-read-backwards==3.1.0
idna==3.7
psutil==5.9.8
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.2
urllib3==2.2.1
